requirements.txt
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.4
markdown2==2.3.1
olefile==0.44
Pillow==4.0.0
